+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-## Unreleased
+## 0.84.0
+
+<!-- release-size: minor -->
 
 ### Added